Split setup.js into env setup, db setup, fix Docker setup (#165)
